Refactored TextView builder to new BooksController class
